requirements.txt
django-allauth == 65.7.0 
folium == 0.19.5 
geopandas == 1.0.1  
ipykernel == 6.29.5 
pip == 24.2   
plotly == 6.0.0  
prophet == 1.1.6  
scikit-learn == 1.6.1
seaborn == 0.13.2
xgboost == 2.1.4